astrocyte-local-py/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "astrocyte-local"
version = "0.1.0"
description = "Zero-infrastructure memory for AI coding agents — Context Tree + FTS5 search"
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.11"
authors = [
    { name = "AstrocyteAI" },
]
dependencies = [
    "astrocyte>=0.2.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
mcp = ["fastmcp>=3.0"]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.4",
]

[project.scripts]
astrocyte-local = "astrocyte_local.cli:main"
astrocyte-local-mcp = "astrocyte_local.mcp:main"

[project.entry-points."astrocyte.engine_providers"]
local = "astrocyte_local:LocalEngineProvider"

[project.urls]
Homepage = "https://github.com/AstrocyteAI/astrocyte-local"
Repository = "https://github.com/AstrocyteAI/astrocyte-local"

[tool.hatch.build.targets.wheel]
packages = ["astrocyte_local"]

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.ruff]
line-length = 120
target-version = "py311"

[tool.ruff.lint]
select = ["E", "W", "F", "I"]

[tool.ruff.format]
quote-style = "double"
